New function: find_nodes_by_name
This function aims to provide users with a better way
to connect to their devices by their names.

Currently the library only supports connecting to
Bluetooth devices by a given node ID. This ID is set
by the library when performing a scan for devices and
also by parsing a text file with pre-set device in-
formation when calling the init_blue() function.

The problem is that right now the users can only see
which ID was allocated to the scanned devices by
printing in the stdout. With this new function users
can call it to find out which IDs were allocated to
devices with a given name and don't relying on the
stdout. This way users can write more autonomous
applications.

Signed-off-by: Matheus Souza <mfbsouza.it@gmail.com>